backend/python_ai_service/requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
langchain-community>=0.0.10
langchain-core>=0.1.7,<0.2.0
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
anthropic>=0.18.0 